Switch 'enabledonly' parameter in metric\manager::get_metrics() to default to true. Move frequency and enabled functions from metric\buitlin to metric\base. Switch order of if statements in collect_metrics_task::get_frequency_cutoff().
